% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abate.R
\name{abate}
\alias{abate}
\title{Average Treatment Effect for A/B Testing}
\usage{
abate(model, treatment)
}
\arguments{
\item{model}{An existing linear regression (\code{lm}) object containing all
control variables and the treatment variable.}

\item{treatment}{The treatment variable name Must be in quotations.}
}
\value{
A \code{flextable} object
}
\description{
This function uses linear regression to calculate the average
     treatment effects both without controls and with controls. The function
     returns a \code{flextable} object.
}
\details{
REQUIRED PACKAGES:
\itemize{
  \item \code{dplyr}
  \item \code{gtsummary}
  \item \code{flextable}
}
}
\examples{
abate(visit, "promotion")
}
