% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acplots.R
\name{acplots}
\alias{acplots}
\title{Easy ACF and PACF Plots}
\usage{
acplots(
  data,
  tvar,
  obs,
  datetype = c("ym", "yq", "yw"),
  h,
  lags = 25,
  d = 0,
  D = c(0, 1)
)
}
\arguments{
\item{data}{The name of the data frame with the time variable and 
measure variable}

\item{tvar}{The variable that identifies the time period. Must be in 
quotations.}

\item{obs}{The variable that identifies the measure. Must be in 
quotations.}

\item{datetype}{Indicates the time period type. Takes one of the following 
    values:
\itemize{
\item \code{"ym"} if the time period is year-month
\item \code{"yq"} if the time period is year-quarter
\item \code{"yw"} if the time period is year-week
}}

\item{h}{An integer indicating the number of holdout/forecast periods.}

\item{lags}{An integer indicating how many lags should be shown on the plot 
(Default: \code{lags=25}).}

\item{d}{An integer indicating the order of non-seasonal differencing
(Default: \code{d=0}).}

\item{D}{An integer indicating the order of seasonal differencing
(Default: \code{D=0}).}
}
\value{
A list containing the following objects:
\itemize{
  \item \code{$acf} contains the ACF plot
  \item \code{$pacf} contains the PACF plot
}
}
\description{
This function creates ACF and PACF plots needed for ARIMA 
    forecasting.
}
\details{
For best results, the results of the function should be saved to an object. \cr
REQUIRED PACKAGES:
\itemize{
  \item \code{fpp3} (NOTE: This package requires several other packages)
}
}
\examples{
out <- acplots(msales, "t", "sales", "ym", 12, lags=25, d=0, D=1)
out$acf
out$pacf
}
