% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advtsales.R
\docType{data}
\name{advtsales}
\alias{advtsales}
\title{Advertising and Sales}
\format{
A data frame with 200 rows and 5 variables:
\describe{
\item{id}{Numeric. Firm identification number.}
\item{ad_tv}{Numeric. Television advertising spend (in 000s).}
\item{ad_radio}{Numeric. Radio advertising spend (in 000s).}
\item{ad_paper}{Numeric. Newspaper advertising spend (in 000s).}
\item{sales}{Numeric. Sales revenue (in millions).}
}
}
\source{
Unknown.
}
\usage{
advtsales
}
\description{
Advertising and sales data for 200 firms.
}
\details{
Used for simple linear regression samples.
}
\keyword{datasets}
