% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asmnl_me.R
\name{asmnl_me}
\alias{asmnl_me}
\title{Easy Alternative Specific MNL Marginal Effects}
\usage{
asmnl_me(mod)
}
\arguments{
\item{mod}{The object containing the results of the \code{asmnl_est} function}
}
\value{
Text output of predicted probabilities for each brand and the
    marginal effects for each independent variable
}
\description{
This function provides marginal effects tables for a saved
    \code{asmnl_est} object
}
\details{
REQUIRED PACKAGES:
\itemize{
  \item mlogit
}
}
\examples{
altspec_model <- asmnl_est(formula=myformula, data=train.yog, id="id",
                           alt="brand", choice="choice", 
                           testdata=test.yog)
asmnl_me(altspec_model)
}
