% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoarima.R
\name{autoarima}
\alias{autoarima}
\title{Easy Automatic ARIMA Modeling}
\usage{
autoarima(
  data,
  tvar,
  obs,
  datetype = c("ym", "yq", "yw"),
  h,
  arima,
  sarima,
  auto = c("Y", "N")
)
}
\arguments{
\item{data}{The name of the data frame with the time variable and 
measure variable}

\item{tvar}{The variable that identifies the time period. Must be in 
quotations.}

\item{obs}{The variable that identifies the measure. Must be in 
quotations.}

\item{datetype}{Indicates the time period type. Takes one of the following 
    values:
\itemize{
\item \code{"ym"} if the time period is year-month
\item \code{"yq"} if the time period is year-quarter
\item \code{"yw"} if the time period is year-week
}}

\item{h}{An integer indicating the number of holdout/forecast periods.}

\item{arima}{Provides the non-seasonal autoregressive (\emph{p}), 
differencing (\emph{d}), and moving average (\emph{q}) parameters, 
in the form of \code{c(p,d,q)}, where the parameters are integers.}

\item{sarima}{Provides the seasonal autoregressive (\emph{P}), 
differencing (\emph{D}), and moving average (\emph{Q}) parameters, 
in the form of \code{c(P,D,Q)}, where the parameters are integers.}

\item{auto}{Instructs the function to do automatic searhing or not. Can be
    one of two options.
\itemize{
\item \code{"Y"} An automatic search is performed to find the “best” model
\item \code{"N"} An automatic search is not performed
}}

\item{lags}{An integer indicating how many lags should be shown on the plot 
(Default: \code{lags=25}).}
}
\value{
A list containing the following objects:
\itemize{
  \item \code{$plot} contains the ARIMA methods plot
  \item \code{$acc} contains the accuracy measures
  \item \code{$fcresplot} contains a plot of the holdout period residuals
  \item \code{$fcresid} contains a dataframe of the holdout period residuals 
      (seldom used separately)
  \item \code{$acresid} contains a plot of the ACF/PACF residuals for the 
      model(s)
  \item \code{$wn} contains cumulative periodogram(s) and results of the 
      Ljung-Box White Noise test(s)
}
}
\description{
This function runs an ARIMA model as specified by the user and 
    an automatic search for the “best” model if requested.
}
\details{
For best results, the results of the function should be saved to an object. \cr
REQUIRED PACKAGES:
\itemize{
  \item \code{fpp3} (NOTE: This package requires several other packages)
  \item \code{ggfortify}
  \item \code{stringr}
  \item \code{cowplot}
}
NOTE 1:  The model names are:
\itemize{
  \item \strong{Self} for the model specified by the user
  \item \strong{Auto} for the "best" model if requested by the user
}
NOTE 2:  If \code{auto="Y"} is specified, this function can take several 
    minutes to run.
}
\examples{
# Running only the model specified by the user
selfonly <- autoarima(msales, "t", "sales", "ym", 12, 
                      arima=c(2,1,0), sarima=c(2,1,0), auto="N")
selfonly$plot
selfonly$acc
selfonly$fcresplot
selfonly$ascresid
selfonly$wn

# Running the automatic search
arima <- autoarima(msales, "t", "sales", "ym", 12, 
                   arima=c(2,1,0), sarima=c(2,1,0), auto="Y")
arima$plot
arima$acc
arima$fcresplot
arima$ascresid
arima$wn
}
