% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builderstores.R
\docType{data}
\name{builderstores}
\alias{builderstores}
\title{Builder Stores}
\format{
A data frame with 280 rows and 8 columns:
\describe{
  \item{\code{brand}}{Factor variable with the name of the brand}
  \item{\code{conv}}{Shopping at location is convenient  
      (1:Strongly Disagree to 6:Strongly Agree)}
  \item{\code{choice}}{Has a large selection of products to choose from  
      (1:Strongly Disagree to 6:Strongly Agree)}
  \item{\code{price}}{The prices are low  
      (1:Strongly Disagree to 6:Strongly Agree)}
  \item{\code{product}}{The products are high quality  
      (1:Strongly Disagree to 6:Strongly Agree)}
  \item{\code{service}}{Provides high quality service  
      (1:Strongly Disagree to 6:Strongly Agree)}
  \item{\code{avail}}{The product I want is usually in-stock 
      (1:Strongly Disagree to 6:Strongly Agree)}
  \item{\code{pref}}{Brand preference rating 
      (1:Not at all preferred to 5:Very preferred)}
}
}
\source{
Unknown
}
\usage{
data(greekbrands)
}
\description{
Attribute and preference data for 4 home improvement stores (names are 
    masked). Each brand is rated 70 times.
}
\keyword{datasets}
