% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cereal.R
\docType{data}
\name{cereal}
\alias{cereal}
\title{Cereal Price and Sales}
\format{
A data frame with 380 rows and 3 variables:
\describe{
\item{week}{Integer. Week number.}
\item{sold}{Integer. Weekly sales (in 000s).}
\item{price}{Numeric. Price in dollars.}
}
}
\source{
Unknown.
}
\usage{
cereal
}
\description{
Price and sales data for one cereal brand over 380 weeks.
}
\details{
Used for modeling price response model.
}
\keyword{datasets}
