% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheese.R
\docType{data}
\name{cheese}
\alias{cheese}
\title{Cheese Price, Advertising, and Sales}
\format{
A data frame with 1863 rows and 5 variables:
\describe{
\item{vol}{Integer. Weekly volume sales.}
\item{price}{Numeric. Price in dollars.}
\item{store}{Character. Store name (KROGER, FOOD LION, WINN DIXIE.}
\item{loc}{Character. City name.}
\item{adv}{Numeric. Index of advertising spending.}
}
}
\source{
Unknown.
}
\usage{
cheese
}
\description{
Price, advertising and sales data for one cheese brand across 20 cities and 3 stores.
}
\details{
Used for modeling price and advertising response model.
}
\keyword{datasets}
