% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dressing.R
\docType{data}
\name{dressing}
\alias{dressing}
\title{Salad Dressing Sales}
\format{
A data frame with 1,365 rows and 5 columns:
\describe{
  \item{\code{id}}{Unique location/week identifier}
  \item{\code{vol}}{Units sold for the week}
  \item{\code{price}}{Price for the week}
  \item{\code{adv}}{Advertising level for the week}
  \item{\code{store}}{Factor variable indicating the retailer (Alb = 
      Albertson's; WD = Winn-Dixie)}
}
}
\source{
Unknown
}
\usage{
data(dressing)
}
\description{
Weekly data on the purchase of salad dressing across five locations for
    two stores over about 2 years. \cr
    \cr
Used in Lab Assignment 9
}
\keyword{datasets}
