% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_mp_plot.R
\name{easy_mp_plot}
\alias{easy_mp_plot}
\title{Quick Marginal Effects Plot}
\usage{
easy_mp_plot(model, focal, int = NULL)
}
\arguments{
\item{model}{A fitted \code{lm} model or binary logistic \code{glm} model
(\code{family = "binomial"}).}

\item{focal}{Character string; name of the focal predictor variable.}

\item{int}{Optional character string; name of the interaction variable.}
}
\value{
A \code{ggplot} object.
}
\description{
Convenience wrapper around \code{\link{easy_mp}} that returns only the
\code{ggplot} object.
}
\examples{
\dontrun{
model1 <- lm(nps ~ age + nflights, data = airlinesat)
easy_mp_plot(model1, "age")
}

}
\seealso{
\code{\link{easy_mp_plot}}
}
\keyword{internal}
