% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emailcampw.R
\docType{data}
\name{email.camp.w}
\alias{email.camp.w}
\title{Email Campaign A/B Testing Data (Women's Merchandise)}
\format{
A data frame with 10,000 rows and 9 variables:
\describe{
\item{visit}{Integer. Site visit within two weeks after the experiment
(\code{1 = yes}, \code{0 = no}).}
\item{spend}{Numeric. Amount spent on the site within two weeks after the experiment.}
\item{recency}{Numeric. Months since last purchase prior to the experiment.}
\item{history}{Numeric. Amount spent on the site in the past year.}
\item{womens}{Integer. Purchased women's merchandise in the past year
(\code{1 = yes}, \code{0 = no}).}
\item{zip}{Factor. Residential area type
(\code{Rural}, \code{Suburban}, \code{Urban}).}
\item{newbie}{Integer. New customer in the past year
(\code{1 = yes}, \code{0 = no}).}
\item{channel}{Factor. Prior purchase channel
(\code{Phone}, \code{Web}, \code{Multichannel}).}
\item{promotion}{Integer. Received the promotion
(\code{1 = yes}, \code{0 = no}).}
}
}
\source{
MineThatData blog (2008).
}
\usage{
email.camp.w
}
\description{
Results of an A/B test of an email marketing campaign for women's merchandise.
}
\keyword{datasets}
