% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fccompare.R
\name{fccompare}
\alias{fccompare}
\title{Forecast Compare}
\usage{
fccompare(results, models)
}
\arguments{
\item{results}{A list of the stored methods results created with the 
\code{list} function}

\item{models}{A vector of the models you want to compare. The names of the
models need to be in quotations and need to match the names used is the
stored methods results.}
}
\value{
\itemize{ 
  \item \code{$acc} Table of accuracy measures
  \item \code{$fcresplot} Plot of holdout period residuals
}
}
\description{
This function compares forecasting models previous ran
}
\details{
REQUIRED PACKAGES:
\itemize{
  \item purrr
  \item dplyr
}
}
\examples{
# Create stored methods
naive <- naivefc(msales, "t", "sales", "ym", 12)
smooth <- smoothfc(msales, "t", "sales", "ym", 12)
linreg <- linregfc(msales, "t", "sales", "ym", 12)

# Create list object with results and vector with method names
results <- list(naive, smooth, linreg)
models <- c("Naive", "Mov.Ave", "Lin.Reg.Trend")

# Compare models
fccompare(results, models)
}
