% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffattrib.R
\docType{data}
\name{ffattrib}
\alias{ffattrib}
\title{Fast Food Attribute Perception Data}
\format{
A data frame with 320 rows and 10 variables:
\describe{
\item{id}{Integer. Respondent identifier.}
\item{rest}{Factor. Fast food restaurant name.}
\item{taste}{Integer. Perceptions of restaurant taste (1-low to 5-high scale).}
\item{speed}{Integer. Perceptions of restaurant speed (1-low to 5-high scale).}
\item{value}{Integer. Perceptions of restaurant value (1-low to 5-high scale).}
\item{variety}{Integer. Perceptions of restaurant menu variety (1-low to 5-high scale).}
\item{clean}{Integer. Perceptions of restaurant cleanliness (1-low to 5-high scale).}
\item{conv}{Integer. Perceptions of restaurant convenience (1-low to 5-high scale).}
\item{healhty}{Integer. Perceptions of restaurant healthiness (1-low to 5-high scale).}
\item{pref}{Integer. Restaurant preference (1-low to 5-high scale).}
}
}
\source{
Data from MKT 3200 classes.
}
\usage{
ffattrib
}
\description{
Attribute perception ratings related to fast food restaurants in Bowling
Green, Ohio. Collected from BGSU students.
}
\keyword{datasets}
