% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffseg.R
\docType{data}
\name{ffseg}
\alias{ffseg}
\title{Fast Food Segmentation Data}
\format{
A data frame with 752 rows and 29 columns:
\describe{
  \item{frequency}{How many times per week they eat fast food (categorical)}
  \item{usertype}{Light vs. Heavy vs. Non}
  \item{spend}{How much typically spent (categorical)}
  \item{eatin}{Prefer eat-in (1) or take-out/delivery (7)}
  \item{clean}{Importance of cleanliness}
  \item{variety}{Importance of menu variety}
  \item{quality}{Importance of food quality}
  \item{location}{Importance of location}
  \item{speed}{Importance of speed of service}
  \item{healthy}{Importance of healthy options availability}
  \item{price}{Importance of price}
  \item{friendly}{Importance of employee friendliness}
  \item{hours}{Importance of hours}
  \item{type}{Type of fast food preferred (categorical)}
  \item{live}{Where they live during the school year (categorical)}
  \item{mealplan}{Whether or not they have a meal plan (yes/no)}
  \item{gender}{Gender}
  \item{class}{Class level (Freshman, Sophmomore, Junior, Senior)}
  \item{college}{BGSU college they are apart of}
  \item{bclean to bhours}{Dichotomous importance of fast food attributes}
  \item{id}{Respondent id number}
}
}
\source{
Data from MKT 3200 classes
}
\usage{
data(ffseg)
}
\description{
Dataset containing attribute importance and respondent demographics from
    BGSU students about eating fast food.
}
\keyword{datasets}
