% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gainlift.R
\name{gainlift}
\alias{gainlift}
\title{Gain and Lift Tables/Charts}
\usage{
gainlift(MOD, TRAIN, TEST, POSITIVE)
}
\arguments{
\item{MOD}{The saved results of a binary logistic regression glm() call}

\item{TRAIN}{The name of the data frame containing the training data}

\item{TEST}{The name of the data frame containing the test/holdout data}

\item{POSITIVE}{The level of the factor variable that is "true" or a
"success"; must be in quotations}
}
\value{
A list containing the following objects:
\itemize{
  \item $gaintable - Gain table for both training and test/holdout
      sample demi-deciles
  \item $gainplot - ggplot object containing the gain plot
  \item $lifttable - Lift table for both training and test/holdout
      sample demi-deciles
  \item $liftplot - ggplot object containing the lift plot
}
}
\description{
This function accepts a binary logistic regression - glm() -
   object and returns a list object of gain and lift tables and charts for
   both training data and test/holdout data.
}
\details{
For best results, the results of the function should be saved to an object. \cr
REQUIRED PACKAGES:
\itemize{
  \item ggplot2
  \item dplyr
  \item tidyr
}
}
\examples{
model <- glm(buy ~ age + gender, train, family="binomial")

# Save gainlift results to object called gl
gl <- gainlift(model, train, test, "Yes")

# Request results from the "gl" object
gl$gaintable
gl$gainplot
gl$lifttable
gl$liftplot
}
