% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greekbrands.R
\docType{data}
\name{greekbrands}
\alias{greekbrands}
\title{Greek Brands}
\format{
A data frame with 1,000 rows and 10 columns:
\describe{
  \item{\code{perform}}{Performance rating (1:low to 10:high)}
  \item{\code{leader}}{Leader in innovation rating (1:low to 10:high)}
  \item{\code{lattech}}{Latest technology rating (1:low to 10:high)}
  \item{\code{fun}}{Fun products rating  (1:low to 10:high)}
  \item{\code{serious}}{Serious products rating  (1:low to 10:high)}
  \item{\code{bargain}}{Bargain prices rating  (1:low to 10:high)}
  \item{\code{value}}{Value for the money rating  (1:low to 10:high)}
  \item{\code{trendy}}{Trendy products rating  (1:low to 10:high)}
  \item{\code{pref}}{Brand preference rating  (1:low to 10:high)}
  \item{\code{brand}}{Factor variable with the name of the brand}
}
}
\source{
Simulated
}
\usage{
data(greekbrands)
}
\description{
Simulated data of attribute ratings and brand preference for 10 fictional
    technology brands. Each brand has 100 ratings.
}
\keyword{datasets}
