% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linregfc.R
\name{linregfc}
\alias{linregfc}
\title{Easy Linear Regression Forecasting}
\usage{
linregfc(data, tvar, obs, datetype = c("ym", "yq", "yw"), h)
}
\arguments{
\item{data}{The name of the data frame with the time variable and 
measure variable}

\item{tvar}{The variable that identifies the time period. Must be in 
quotations.}

\item{obs}{The variable that identifies the measure. Must be in 
quotations.}

\item{datetype}{Indicates the time period type. Takes one of the following 
    values:
\itemize{
\item \code{"ym"} if the time period is year-month
\item \code{"yq"} if the time period is year-quarter
\item \code{"yw"} if the time period is year-week
}}

\item{h}{An integer indicating the number of holdout/forecast periods.}
}
\value{
A list containing the following objects:
\itemize{
  \item \code{$plot} contains the smoothing methods plot
  \item \code{$acc} contains the accuracy measures
  \item \code{$fcresplot} contains a plot of the holdout period residuals
  \item \code{$fcresid} contains a dataframe of the holdout period residuals 
      (seldom used separately)
}
}
\description{
This function analyzed time series data using linear regression 
    forecasting methods.
}
\details{
For best results, the results of the function should be saved to an object. \cr
REQUIRED PACKAGES:
\itemize{
  \item \code{fpp3} (NOTE: This package requires several other packages)
}
NOTE:  The model names are:
\itemize{
  \item \strong{Lin.Reg.Trend} for the linear regression with trend model
  \item \strong{Lin.Reg.Seas.Trend} for the linear regression with trend 
      and seasonality model
}
}
\examples{
linreg <- linregfc(msales, "t", "sales", "ym", 12)
linreg$plot
linreg$acc
linreg$fcresplot
}
