% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_beta.R
\name{lm_beta}
\alias{lm_beta}
\title{Standardized Beta Coefficients}
\usage{
lm_beta(object, digits = 8)
}
\arguments{
\item{object}{The saved results of a lm() call}

\item{digits}{The number of digits to display (default = 8)}
}
\value{
The standardized beta coefficients
}
\description{
This function accepts a lm() object and returns the
   standardized beta coefficients. Code adapted from the lm.beta package
}
\examples{
model <- lm(nps ~ age + nflights, data=airlinesat)

lm_beta(model, 4)
}
