% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logreg_cut.R
\name{logreg_cut}
\alias{logreg_cut}
\title{Sensitivity/Specificity Plot}
\usage{
logreg_cut(MOD, DATA, POS)
}
\arguments{
\item{MOD}{The saved results of a binary logistic regression glm() call}

\item{DATA}{The name of the data frame for which the classification matrix
should be produced (i.e., training data, holdout data, etc.)}

\item{POSITIVE}{The level of the factor variable that is "true" or a
"success"; must be in quotations}
}
\value{
A ggplot object
}
\description{
This function accepts a binary logistic regression - glm() -
   object and returns a sensitivity/specificity plot.
}
\details{
REQUIRED PACKAGES:
\itemize{
  \item ggplot2
}
}
\examples{
model <- glm(buy ~ age + gender, train, family="binomial")

logreg_cut(model, train, "Yes")
}
