% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logreg_roc.R
\name{logreg_roc}
\alias{logreg_roc}
\title{ROC Curve}
\usage{
logreg_roc(MOD, DATA)
}
\arguments{
\item{MOD}{The saved results of a binary logistic regression glm() call}

\item{DATA}{The name of the data frame for which the classification matrix
should be produced (i.e., training data, holdout data, etc.)}
}
\value{
ggplot2 object
}
\description{
This function accepts a binary logistic regression - glm() -
   object and returns the ROC curve.
}
\details{
REQUIRED PACKAGES:
\itemize{
  \item pROC
  \item ggplot2
}
}
\examples{
model <- glm(buy ~ age + gender, train, family="binomial")

logreg_roc(model, train)
}
