% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msales_qsales.R
\docType{data}
\name{msales}
\alias{msales}
\alias{qsales}
\title{Monthly Sales Data (Legacy)}
\format{
\describe{
\item{msales}{A data frame.}
\item{qsales}{A data frame.}
}

An object of class \code{data.frame} with 92 rows and 2 columns.
}
\usage{
msales

qsales
}
\description{
Legacy datasets used in older course materials.
}
\keyword{datasets}
\keyword{internal}
