% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhc.R
\name{myhc}
\alias{myhc}
\title{Easy Hierarchical Agglomerative Clustering}
\usage{
myhc(
  data,
  dist = c("euc", "euc2", "max", "abs", "bin"),
  method = c("ward", "single", "complete", "average"),
  cuts,
  clustop = c("N", "Y")
)
}
\arguments{
\item{data}{A data frame containing only the variables on which to cluster}

\item{dist}{Distance measure to be used for the similarity/dissimilarity
    matrix. Must be in quotations. Takes one of the following values:
\itemize{
  \item \code{dist="euc"} provides Euclidean distance
  \item \code{dist="euc2"} provides Euclidean Squared distance
  \item \code{dist="max"} provides Maximum distance
  \item \code{dist="abs"} provides Absolute distance
  \item \code{dist="bin"} provides a distance measure for a set of binary-only variables
}}

\item{method}{Indicates the linkage method. Must be in quotations. Takes one
    of the following values:
\itemize{
\item \code{method="single"} provides Single linkage
\item \code{method="complete"} provides Complete linkage
\item \code{method="average"} provides Average linkage
\item \code{method="ward"} provides Ward’s linkage
}}

\item{cuts}{Indicates how many clusters are desired for the solution (optional).}

\item{clustop}{Indicates if stopping indices are wanted. Must be in quotations.
    Takes one of the following values:
\itemize{
\item \code{clustop="Y"} if indices are wanted
\item \code{clustop="N"} if indices are not wanted
}}
}
\value{
A list containing the following objects:
\itemize{
  \item if \code{cuts} are provided:
      \itemize{
         \item Dendrogram of the top \emph{n} branches, where \emph{n} is
             the highest number of clusters provided by \code{cuts}.
             Displayed automatically and not saved to the list object.
         \item \code{$kcount} is a table of cluster sizes
         \item \code{$kperc} is table of cluster size percentages
         \item \code{$hc} is the \code{hclust} object created
         \item \code{$stop} is a table of stopping indices for 1 to 10 clusters
              if \code{clustop="Y"}
      }
  \item if \code{cuts} are not provided:
      \itemize{
         \item Dendrogram with all branches.  Displayed automatically and
             not saved to the list object.
         \item \code{$stop} is a table of stopping indices for 1 to 10 clusters
              if \code{clustop="Y"}
      }
}
}
\description{
This function performs hierarchical agglomerative clustering,
    allowing the user to pick a specific number of clusters, get stopping
    indices, etc.
}
\details{
For best results, the results of the function should be saved to an object. \cr
REQUIRED PACKAGES:
\itemize{
  \item dendextend
  \item dplyr
  \item NbClust
}
}
\examples{
#No cuts with stopping indices
eg1 <- myhc(sc.clvar, "euc", "ward", "Y")
eg1$stop

#One cut without stopping indices
eg2 <- myhc(sc.clvar, "max", "ward", cuts=5, "N")
eg2$kcount
eg2$kperc

#Multiple cuts without stopping indices
eg3 <- myhc(sc.clvar, "abs", "ward", cuts=c(2,3,4,6), "N")
eg3$kcount
eg3$kperc
}
