% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymp.R
\name{mymp}
\alias{mymp}
\title{Easy Margin Plots}
\usage{
mymp(model, focal, int = NULL)
}
\arguments{
\item{model}{A saved linear regression (lm) model or binary logistic
regression (glm, family="binomial") model.}

\item{focal}{Name of the focal predictor variable. Must be in quotations.}

\item{int}{Name of the interaction predictor variable. Must be in quotations. 
Can be excluded if focal variable only wanted (or no interaction exists).}
}
\value{
\itemize{ 
  \item \code{$plot} A ggplot object.
  \item \code{$ptable} A ggeffects table that can be saved as a data frame
   object.
}
}
\description{
This function produces margin plots for a focal predictor or a
   focal predictor interaction with a another predictor from a linear 
   or binary logistic regression
}
\details{
REQUIRED PACKAGES:
\itemize{
  \item ggplot2
  \item dplyr
  \item ggeffects
  \item insight
}
}
\examples{
#No interaction
data(airlinesat)
model1 <- lm(nps ~ age + nflights, data=airlinesat)
mymp(model, "age")

#Interaction
data(directmktg)
model2 <- glm(buy ~ salary + age*gender, data=directmktg, family="binomial")
mymp(model2, "age", "gender")
}
