% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/or_table.R
\name{or_table}
\alias{or_table}
\title{Odds Ratio Coefficients Table}
\usage{
or_table(MOD, DIGITS = 4, LEVEL = 95)
}
\arguments{
\item{MOD}{The saved results of a binary logistic regression glm() call}

\item{DIGITS}{The number of digits to display (default = 4)}

\item{LEVEL}{Confidence level for CI (default = 95)}
}
\value{
Table of Odds Ratio Coefficients
}
\description{
This function accepts a binary logistic regression - glm() -
   object and returns the odds ratio coefficients, p-values, and confidence
   intervals.
}
\examples{
prelim <- glm(buy ~ age + gender, directmktg, family="binomial")

or_table(prelim)
or_table(prelim, 3, 90)
}
