% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percmap.R
\name{percmap}
\alias{percmap}
\title{Easy Perceptual and Joint Space Maps from PCA}
\usage{
percmap(data, group, pref)
}
\arguments{
\item{data}{A data frame containing the variables on which to perform the
principal components analysis. Can include a preference variable.}

\item{group}{Name of the grouping variable. Must be in quotations.}

\item{pref}{Name of the preference variable. Must be in quotations. Can be 
excluded if no preference variable exists.}
}
\value{
\itemize{ 
  \item If a \code{pref} variable is not included, returns a perceptual map
  \item If a \code{pref} variable is included, returns a joint space map
}
}
\description{
This function produces perceptual or joint spaced maps based
    on the first two principal components of a PCA
}
\details{
REQUIRED PACKAGES:
\itemize{
  \item ggplot2
  \item dplyr
}
}
\examples{
#Perceptual map
pmdata <- greekbrands \%>\%
   select(perform, leader, fun, serious bargain, value, brand)
percmap(pmdata, group="brand")

#Joint Space map
jsmdata <- greekbrands \%>\%
   select(perform, leader, fun, serious bargain, value, pref, brand)
percmap(pmdata, group="brand", pref="pref")
}
