% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcheck.R
\name{rcheck}
\alias{rcheck}
\title{Randomization check for A/B Testing}
\usage{
rcheck(data, treatment, outcome = NULL, nice = c("no", "ft", "ht"))
}
\arguments{
\item{data}{The name of the data frame containing the treatment variable and
the covariates.}

\item{treatment}{The variable name identifying the treatment variable. Must 
be in quotations.}

\item{outcome}{The name or names of the variables that identifies the 
outcome variables. Default is \code{NULL}. Must be in quotations.}

\item{nice}{Format for the output
\itemize{
\item \code{"no"} for standard output
\item \code{"ft"} for output using package \code{flextable}
\item \code{"ht"} for output using package \code{htmlTable}
}}
}
\value{
A table containing the results of the randomization ehck
}
\description{
This function checks if the characteristics/covariates used for
    uplift modeling from an A/B Test were randomly assigned to the test and
    control groups.
}
\details{
REQUIRED PACKAGES:
\itemize{
  \item \code{fastDummies}
  \item \code{htmlTable} (if \code{nice="ht"} option used)
  \item \code{flextable} (if \code{nice="ft"} option used)
}
}
\examples{
rcheck(email.camp.w, "promotion", c("visit", "spend"), nice="ft")
}
