% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales03.R
\docType{data}
\name{sales03}
\alias{sales03}
\title{Durable Manufacturer Sales}
\format{
\itemize{
   \item \code{sales03} contains data for 120 months (120 rows and 2 columns)
}
\describe{
  \item{\code{t}}{Time period}
  \item{\code{sales}}{Monthly sales}
}
}
\source{
Unknown
}
\usage{
data(sales03)
}
\description{
Monthly sales data (in $000s) for a durable goods manufacturer from January 2008 through December 2017.
}
\keyword{datasets}
