% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothfc.R
\name{smoothfc}
\alias{smoothfc}
\title{Easy Smoothing Forecasting}
\usage{
smoothfc(data, tvar, obs, datetype = c("ym", "yq", "yw"), h)
}
\arguments{
\item{data}{The name of the data frame with the time variable and 
measure variable}

\item{tvar}{The variable that identifies the time period. Must be in 
quotations.}

\item{obs}{The variable that identifies the measure. Must be in 
quotations.}

\item{datetype}{Indicates the time period type. Takes one of the following 
    values:
\itemize{
\item \code{"ym"} if the time period is year-month
\item \code{"yq"} if the time period is year-quarter
\item \code{"yw"} if the time period is year-week
}}

\item{h}{An integer indicating the number of holdout/forecast periods.}
}
\value{
A list containing the following objects:
\itemize{
  \item \code{$plot} contains the smoothing methods plot
  \item \code{$acc} contains the accuracy measures
  \item \code{$fcresplot} contains a plot of the holdout period residuals
  \item \code{$fcresid} contains a dataframe of the holdout period residuals 
      (seldom used separately)
}
}
\description{
This function analyzed time series data using smoothing 
    forecasting methods.
}
\details{
For best results, the results of the function should be saved to an object. \cr
REQUIRED PACKAGES:
\itemize{
  \item \code{fpp3} (NOTE: This package requires several other packages)
  \item \code{slider}
}
NOTE:  The model names are:
\itemize{
  \item \strong{Move.Ave} for the moving average model
  \item \strong{Exp.Smooth} for the exponential smoothing model
  \item \strong{H-W.Add} for the Holt-Winters Additive model
  \item \strong{H-W.Mult} for the Holt-Winters Multiplicative modelo
}
}
\examples{
smooth <- smoothfc(msales, "t", "sales", "ym", 12)
smooth$plot
smooth$acc
smooth$fcresplot
}
