% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stmnl_cm.R
\name{stmnl_cm}
\alias{stmnl_cm}
\title{Classification Matrix and Statistics for Standard MNL}
\usage{
stmnl_cm(model, data)
}
\arguments{
\item{model}{A \code{multinom} object from the \code{nnet} package}

\item{data}{The name of the data frame for which the classification matrix
should be produced (i.e., training data, holdout data, etc.)}
}
\value{
Classification matrix and statistics
}
\description{
This function accepts a multinomial logistic regression object
   from the \code{multinom} function of the \code{nnet} package and returns 
   the classification matrix and associated statistics.
}
\examples{
stmnl_cm(model, train)
}
