% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stmnl_pp.R
\name{stmnl_pp}
\alias{stmnl_pp}
\title{Average Predicted Probabilities for Standard MNL}
\usage{
stmnl_pp(model, focal, xlab = NULL)
}
\arguments{
\item{model}{A \code{multinom} object from the \code{nnet} package}

\item{focal}{The name of the variable (in quotes) for which predicted 
probabilities are wanted}

\item{xlab}{The desired x-axis label for the plot, in quotes; optional}
}
\value{
Predicted probabilities table ($table) and ggplot ($plot). Can be
    saved as an object.
}
\description{
This function accepts a multinomial logistic regression object
   from the \code{multinom} function of the \code{nnet} package and returns 
   a table and plot of average predicted probabilities.
}
\details{
For continuous focal variables, the table produced uses three
    levels to calculate predicted probabilities: -1 SD, Mean, +1 SD \cr
    \cr
REQUIRED PACKAGES:
\itemize{
  \item tidyr
  \item effects
  \item dplyr
  \item ggplot2
}
}
\examples{
stmnl_pp(model, "Age", "Age in Years")
}
