% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teleko.R
\docType{data}
\name{teleko}
\alias{teleko}
\title{A/B Testing Data}
\format{
A data frame with 10000 rows and 12 columns:
\describe{
  \item{\code{renew}}{Yes if customer renewed, No otherwise}
  \item{\code{clv}}{predicted revenue in future years}
  \item{\code{promotion}}{Yes if customer received promotion, No otherwise}
  \item{\code{tenure}}{Months since contract started}
  \item{\code{past60}}{Expenditures in 60 days prior to renewal decision}
  \item{\code{past30}}{Expenditures in 30 days prior to renewal decision}
  \item{\code{past7}}{Expenditures in 7 days prior to renewal decision}
  \item{\code{data}}{Yes if customer has data package, No otherwise}
  \item{\code{inter}}{Yes if customer has international plan, No otherwise}
  \item{\code{family}}{Yes if customer has family plan, No otherwise}
  \item{\code{csvc}}{Number of customer service calls in prior 30 days}
  \item{\code{os}}{Phone operating system (Android, IOs, Other)}
}
}
\source{
Harvard Business School Publishing, Case 523-005
}
\usage{
data(teleko)
}
\description{
Results of an A/B test for Teleko Mobile.
}
\keyword{datasets}
