% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.yog.R
\docType{data}
\name{test.yog}
\alias{test.yog}
\title{Yogurt Choice Data (Test)}
\format{
\itemize{
   \item \code{train.yog} contains data for 1,811 choice occassions in long 
       format (7244 rows and 6 columns)
   \item \code{test.yog} contains data for 601 choice occassions in long
       format (2404 rows and 6 columns)
}
\describe{
  \item{\code{id}}{Unique customer identification number}
  \item{\code{feat}}{\code{1} if brand was on feature, \code{0} otherwise}
  \item{\code{price}}{Price of brand per ounce}
  \item{\code{choice}}{\code{Yes} if brand was chosen, \code{No} otherwise}
  \item{\code{brand}}{Brand identifier}
  \item{\code{income}}{Consumer income in 000s}
}
}
\source{
Adapted from mlogit (version 0.3-0) data set
}
\usage{
data(test.yog)
}
\description{
Two data frames taken from 2,412 choice observations of four yogurt brands.
    Original data split into training data (train.yog) and test/holdout
    data (test.yog)
}
\keyword{datasets}
