% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradca.R
\name{tradca}
\alias{tradca}
\title{Traditional Conjoint Results}
\usage{
tradca(formula, data, idvar)
}
\arguments{
\item{formula}{An object with a saved formula with a DV on the left side
and IVs on the right side separated by a tilde (e.g., \code{myform = 
    value ~ attr1 + attr2 + attr3})}

\item{data}{The name of the data frame with the value ratings for each
profile}

\item{idvar}{The variable that identifies the case (in quotations)}
}
\value{
A list containing the following objects:
\itemize{
  \item \code{$pwplot} contains a plot with average part-worth values for
      each attribute
  \item \code{$impplot} contains a plot with the average importances
  \item \code{$casetable} contains a data frame with the part-worths and
      importances for each case
}
}
\description{
This function produces overall part-worth plots for each 
    attribute in a traditional conjoint analysis, along with an attribute
    importance plot and a table with case-level part worths and importances.
}
\details{
For best results, the results of the function should be saved to an object. \cr
REQUIRED PACKAGES:
\itemize{
  \item broom
  \item dplyr
  \item stringr
  \item ggplot2
}
}
\examples{
caform <- value ~ airline + connect + price
results <- tradca(formula = caform, data = airlineca, idvar="caseid")

results$pwplot
results$impplot
results$casetable
}
