% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsplot.R
\name{tsplot}
\alias{tsplot}
\title{Easy Time Series Plot}
\usage{
tsplot(data, tvar, obs, datetype = c("ym", "yq", "yw"), h)
}
\arguments{
\item{data}{The name of the data frame with the time variable and 
measure variable}

\item{tvar}{The variable that identifies the time period. Must be in 
quotations.}

\item{obs}{The variable that identifies the measure. Must be in 
quotations.}

\item{datetype}{Indicates the time period type. Takes one of the following 
    values:
\itemize{
\item \code{"ym"} if the time period is year-month
\item \code{"yq"} if the time period is year-quarter
\item \code{"yw"} if the time period is year-week
}}

\item{h}{An integer indicating the number of holdout/forecast periods.}
}
\value{
A time series plot created in \code{ggplot}
}
\description{
This function creates a time series plot of the data
}
\details{
REQUIRED PACKAGES:
\itemize{
  \item \code{fpp3} (NOTE: This package requires several other packages)
}
}
\examples{
tsplot(qsales, "t", "sales", "yq", 8)
}
