% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tswh.noise.R
\name{tswh.noise}
\alias{tswh.noise}
\title{Easy White Noise Plot and Test}
\usage{
tswh.noise(
  data,
  tvar,
  obs,
  datetype = c("ym", "yq", "yw"),
  h = 0,
  arima,
  sarima
)
}
\arguments{
\item{data}{The name of the data frame with the time variable and 
measure variable}

\item{tvar}{The variable that identifies the time period. Must be in 
quotations.}

\item{obs}{The variable that identifies the measure. Must be in 
quotations.}

\item{datetype}{Indicates the time period type. Takes one of the following 
    values:
\itemize{
\item \code{"ym"} if the time period is year-month
\item \code{"yq"} if the time period is year-quarter
\item \code{"yw"} if the time period is year-week
}}

\item{h}{An integer indicating the number of holdout/forecast periods.}

\item{arima}{Provides the non-seasonal autoregressive (\emph{p}), 
differencing (\emph{d}), and moving average (\emph{q}) parameters, 
in the form of \code{c(p,d,q)}, where the parameters are integers.}

\item{sarima}{Provides the seasonal autoregressive (\emph{P}), 
differencing (\emph{D}), and moving average (\emph{Q}) parameters, 
in the form of \code{c(P,D,Q)}, where the parameters are integers.}
}
\value{
A cumulative periodogram and results of the Ljung-Box White Noise 
    test.
}
\description{
This function creates a cumulative periodogram and white 
    noise test needed for ARIMA forecasting.
}
\details{
REQUIRED PACKAGES:
\itemize{
  \item \code{fpp3} (NOTE: This package requires several other packages)
  \item \code{ggfortify}
}
}
\examples{
tswh.noise(msales, "t", "sales", "ym", 12, arima=c(0,0,0), sarima=c(0,0,0))
tswh.noise(msales, "t", "sales", "ym", 12, c(2,1,0), c(2,1,0))
}
