% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitamins.R
\docType{data}
\name{vitamins}
\alias{vitamins}
\title{Vitamins Preference Data}
\format{
A data frame with 735 rows and 4 columns
\describe{
  \item{\code{brand}}{Consumer brand preference (Alphagenix, Deltastation,
      Sigmastop)}
  \item{\code{female}}{\code{Yes} if consumer is female, \code{No} 
      otherwise}
  \item{\code{age}}{Consumer actual age)}
  \item{\code{income}}{Consumer income level (Low, Medium, High)}
}
}
\source{
Unknown
}
\usage{
data(vitamins)
}
\description{
Preference data for vitamin brands with consumer demographic information.
}
\keyword{datasets}
