% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wssplot.R
\name{wssplot}
\alias{wssplot}
\title{Scree Plot Before k-Means Clustering}
\usage{
wssplot(data, nc = 15, seed = 4320)
}
\arguments{
\item{data}{A data frame containing only the variables on which to cluster}

\item{nc}{An integer representing the maximum number of clusters to plot
(default = 15)}

\item{seed}{A random number seed for reproducible results (default = 4320)}
}
\value{
A ggplot object
}
\description{
This function produces a scree plot for 1 to \emph{n} clusters
}
\details{
REQUIRED PACKAGES:
\itemize{
  \item ggplot2
}
}
\examples{
#Scree plot with default values
wssplot(sc.clvar)

#Scree plot for up to 10 clusters and 1000 for random number seed
wssplot(sc.clvar, 10, 1000)
}
