% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asmnl_est.R
\name{asmnl_est}
\alias{asmnl_est}
\title{Easy Alternative Specific MNL}
\usage{
asmnl_est(formula, data, id, alt, choice, testdata)
}
\arguments{
\item{formula}{An object with a saved formula. The formula is represented by
a DV on the left side separated from the IVs on the right side by a
tilde (\code{~}). The IVs are further separated by having choice-specific 
IVs on the left and the case-specific IVs on the right, separated by a 
vertical line (\code{|}).}

\item{data}{A data frame containing the training data choice observations 
in long format}

\item{id}{Variable that identifies the case (in quotes)}

\item{alt}{Variable that identifies the choice alternative (in quotes)}

\item{choice}{Variable that identifies the if \code{alt} was selected 
(in quotes)}

\item{testdata}{A data frame containing the test/holdout data choice observations 
in long format}
}
\value{
Text output of model fit, odds ratio estimation results, and 
    classification matrices for the training and holdout data.
}
\description{
This function performs alternative specific MNL using the 
    \code{mlogit} package and provides commonly requested results.
}
\details{
For best results, the results of the function should be saved to an object
    to be used in subsequent analyses. \cr
REQUIRED PACKAGES:
\itemize{
  \item mlogit
  \item broom
}
}
\examples{
myformula <- choice ~ feat + price | income
altspec_model <- asmnl_est(formula=myformula, data=train.yog, id="id",
                           alt="brand", choice="choice", 
                           testdata=test.yog)
}
