% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asmnl_mp.R
\name{asmnl_mp}
\alias{asmnl_mp}
\title{Easy Alternative Specific MNL Margin Plots}
\usage{
asmnl_mp(mod, focal, type = c("C", "D"))
}
\arguments{
\item{mod}{The object containing the results of the \code{asmnl_est} function}

\item{focal}{The case-specific IV for which a margin plot is wanted in 
quotations}

\item{type}{The type of the case-specific IV in quotations ("C" for 
continuous; "D" for dummy variable)}
}
\value{
Margin plot using ggplot
}
\description{
This function creates a margin plot for a case-specific 
    independent variable from an estimated alternative specific MNL using
    the \code{asmnl_est} function.
}
\details{
REQUIRED PACKAGES:
\itemize{
  \item mlogit
  \item ggplot2
}
}
\examples{
altspec_model <- asmnl_est(formula=myformula, data=train.yog, id="id",
                           alt="brand", choice="choice", 
                           testdata=test.yog)
asmnl_mp(altspec_model, "income", "C")
}
