% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfast.R
\docType{data}
\name{bfast}
\alias{bfast}
\title{Breakfast Food Preference Data}
\format{
A data frame with 800 rows and 6 columns
\describe{
  \item{\code{agecat}}{Consumer age category (Under 31; 31-45; 46-60; 
      Over 60)}
  \item{\code{gender}}{Consumer gender (Female; Male)}
  \item{\code{marital}}{Consumer marital status (Married; Unmarried)}
  \item{\code{bfast}}{Consumer breakfast food preference (Cereal; Bar; Oatmeal)}
  \item{\code{lifestyle}}{Consumer lifestyle (Active; Inactive)}
  \item{\code{age}}{Consumer actual age}
}
}
\source{
Adapted from IBM SPSS Statistics Sample Files
}
\usage{
data(bfast)
}
\description{
Simulated data for a poll of 880 people about their breakfast preferences, 
    also noting demographics and lifestyle.
}
\keyword{datasets}
