% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cademo.R
\name{cademo}
\alias{cademo}
\title{Demographic Analysis of Traditional Conjoint Results}
\usage{
cademo(formula, data, vars)
}
\arguments{
\item{formula}{An object with a saved formula used in a previous traditional
conjoint analysis that was analyzed with the \code{tradca()} function}

\item{data}{The name of the data frame with the case-level part-worths and
importances that has been appended with demographic variables}

\item{vars}{A concantenated list of the demographics variables to analyze (
in quotations)}
}
\value{
Text correlation matrices and/or regression results
}
\description{
This function produces correlation matrices and importance
    regressions from a data frame of case-level part-worths/importances that
    has been appended with demographics.
}
\details{
To use this function, the \code{$casetable} from the call to the 
    \code{tradca()} function must be appended with the demographic variables
    for each case.  Refer to the course tutorial for code to accomplish
    this.  \cr
     \cr
REQUIRED PACKAGES:
\itemize{
  \item Hmisc
}
}
\examples{
caform <- value ~ airline + connect + price
results <- tradca(formula = caform, data = airlineca, idvar="caseid")
demos <- airlineca \%>\%   
    group_by(caseid) \%>\% 
    summarise(inc=first(inc), type=first(type)) \%>\% 
    bind_cols(.,results$casetable[,2:ncol(results$casetable)])

cademo(caform, demos, c("inc", "type"))
}
