% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capred.R
\name{capred}
\alias{capred}
\title{Traditional Conjoint Analysis Profile Comparison}
\usage{
capred(formula, data, prof1, prof2)
}
\arguments{
\item{formula}{An object with a saved formula used in a previous traditional
conjoint analysis that was analyzed with the \code{tradca()} function}

\item{data}{The name of the data frame with the case-level part-worths and
importances that has been appended with demographic variables}

\item{prof1}{The first profile}

\item{prof2}{The second profile}
}
\value{
Mean utility and 95 percent confidence intervals for each profile
}
\description{
This function compares two profiles (specified by the user) 
    from a traditional conjoint analysis.
}
\details{
To use this function, it works best to create the profiles by saving them 
    to separate objects. The attribute levels for the profiles should be in 
    the form of: \code{attribute_levelnumber}. \cr
    For example, \emph{Delta} is the first level of the \code{airline} 
    attribute, so we would use \code{airline_1} to indicate that level of
    that attribute.
}
\examples{
caform <- value ~ airline + connect + price
results <- tradca(formula = caform, data = airlineca, idvar="caseid")

prof1 <- c("airline_1", "connect_2", "price_3")
prof2 <- c("airline_3", "connect_2", "price_2")

capred(caform, results$casetable, prof1, prof2)
}
