% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directmktg.R
\docType{data}
\name{directmktg}
\alias{directmktg}
\title{Direct Marketing Data}
\format{
A data frame with 400 rows and 5 columns:
\describe{
  \item{userid}{Unique identifier}
  \item{age}{Prospect's age}
  \item{buy}{Buying decision (Yes or No)}
  \item{gender}{Prospect's gender}
  \item{salary}{Prospect's salary (in 000s)}
}
}
\source{
Unknown
}
\usage{
data(directmktg)
}
\description{
Dataset containing purchase decisions for a direct marketing campaign.
}
\keyword{datasets}
