% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greekbrands.R
\docType{data}
\name{greekbrands}
\alias{greekbrands}
\title{Greek Brands}
\format{
A data frame with 1,000 rows and 10 variables:
\describe{
\item{perform}{Integer. Performance rating (1 = low, 10 = high).}
\item{leader}{Integer. Innovation leadership rating (1 = low, 10 = high).}
\item{lattech}{Integer. Latest technology rating (1 = low, 10 = high).}
\item{fun}{Integer. Fun products rating (1 = low, 10 = high).}
\item{serious}{Integer. Serious products rating (1 = low, 10 = high).}
\item{bargain}{Integer. Bargain pricing rating (1 = low, 10 = high).}
\item{value}{Integer. Value-for-money rating (1 = low, 10 = high).}
\item{trendy}{Integer. Trendiness rating (1 = low, 10 = high).}
\item{pref}{Integer. Overall brand preference (1 = low, 10 = high).}
\item{brand}{Factor. Fictional brand identifier.}
}
}
\source{
Simulated.
}
\usage{
greekbrands
}
\description{
Simulated attribute ratings and brand preference data for ten fictional
technology brands.
}
\keyword{datasets}
