% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcenters.R
\name{kcenters}
\alias{kcenters}
\title{k-Means Cluster Centers}
\usage{
kcenters(kobject)
}
\arguments{
\item{kobject}{A saved \emph{k}-means object}
}
\value{
A list containing the following objects:
\itemize{
  \item \code{$table} A table of the cluster center means.
  \item \code{$plot} A plot of the cluster center means.
}
}
\description{
This function take a \emph{k}-means object and creates a table and
    plot of the cluster centers.
}
\details{
For best results, the results of the function should be saved to an object. \cr
REQUIRED PACKAGES:
\itemize{
  \item ggplot2
}
}
\examples{
kmout <- kcenters(k4)
kmout$table
kmout$plot
}
