% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksize.R
\name{ksize}
\alias{ksize}
\title{Cluster Sizes for k-Means Clustering}
\usage{
ksize(data, centers, nstart = 25, seed = 4320)
}
\arguments{
\item{data}{A data frame containing only the variables on which to cluster}

\item{centers}{One or more integers for the specific \emph{k} cluster
solutions to be examined}

\item{nstart}{An integer value for the number of random starting points
to try (default = 25)}

\item{seed}{A random number seed for reproducible results (default = 4320)}
}
\value{
A list containing the following objects:
\itemize{
  \item \code{$kcount} is a table of cluster sizes
  \item \code{$kperc} is a table of cluster size percentages
}
}
\description{
This function produces a produces a count and proportion tables
     for the selected \emph{k} solutions
}
\details{
For easiest use, the results of the function should be saved to an object.
}
\examples{
#Examining three different cluster solutions for default nstart and seed
ks <- ksize(sc.clvar, centers=c(3,4,5))
}
