% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftplot.R
\name{liftplot}
\alias{liftplot}
\title{Easy Lift Plot for Uplift Modeling}
\usage{
liftplot(
  data,
  var = NULL,
  byvar = NULL,
  ar = NULL,
  ci = c(0.9, 0.95, 0.975, 0.99, 0)
)
}
\arguments{
\item{data}{The name of the data frame with the results of an uplift modeling
analysis.}

\item{var}{The variable name for which the error-bars should be created. Must 
be in quotations. Default is \code{"NULL"} for a histogram.}

\item{byvar}{The variable that identifies second variable if side-by-side 
error-bar plots are desired. Must be in quotations. Default is
 \code{"NULL"}.}

\item{ar}{The aspect ratio for the plot. Used if a certain aspect ratio is
desired for presentation purposes. Default is \code{"NULL"}.}

\item{ci}{The type of error-bar desired. Ignored if \code{"var"} is null.
\itemize{
\item \code{"0"} for error-bars to represent 1 standard deviation
\item \code{"c(0.90, 0.95, 0.975, 0.99)"} for error-bars to represent the 
    desired confidence interval.
}}
}
\value{
A ggplot object
}
\description{
This function creates a lift plot following uplift modeling. It
    can create a histogram (if \code{"var"} is null) or an error-bar plot. 
    For continuous variables, it will create an error-bar for the quintile 
    values of the variable. For factor variables, it will create an error-bar 
    for each level of the factor. It can also create side-by-side error-bar 
    plots for two variables simultaneously by using the \code{"byvar"} option.
}
\details{
REQUIRED PACKAGES:
\itemize{
  \item \code{ggplot1}
}
}
\examples{
liftplot(results.visit.w$all, ar=1)
liftplot(results.visit.w$all, "recency", ar=1, ci=0.95)
liftplot(results.visit.w$all, "recency", "history", ar=1, ci=0.95)
}
