% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logreg_cm.R
\name{logreg_cm}
\alias{logreg_cm}
\title{Classification Matrix and Statistics}
\usage{
logreg_cm(MOD, DATA, POSITIVE, CUTOFF = 0.5)
}
\arguments{
\item{MOD}{The saved results of a binary logistic regression glm() call}

\item{DATA}{The name of the data frame for which the classification matrix
should be produced (i.e., training data, holdout data, etc.)}

\item{POSITIVE}{The level of the factor variable that is "true" or a
"success"; must be in quotations}

\item{CUTOFF}{Cutoff threshold for predicted success/failure (default = 0.5)}
}
\value{
Classification matrix and statistics
}
\description{
This function accepts a binary logistic regression - glm() -
   object and returns the classification matrix and associated statistics.
}
\details{
REQUIRED PACKAGES:
\itemize{
  \item caret
}
}
\examples{
model <- glm(buy ~ age + gender, train, family="binomial")

logreg_cm(model, train, "Yes")
logreg_cm(model, train, "Yes", 0.6)
}
