% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaex.R
\name{pcaex}
\alias{pcaex}
\title{Easy PCA}
\usage{
pcaex(data, group, pref, comp = NULL)
}
\arguments{
\item{data}{A data frame containing the variables on which to perform the
principal components analysis. Can include a preference variable.}

\item{group}{Name of the grouping variable. Must be in quotations. Can be
excluded if no grouping variable exists. When provided, the PCA is 
performed on an aggregatded data frame (i.e., mean values by group)}

\item{pref}{Name of the preference variable. Must be in quotations. Can be 
excluded if no preference variable exists.}

\item{comp}{An integer representing the number of components to retain 
(default = NULL, which retains all components).}
}
\value{
A list containing the following objects:
\itemize{
  \item If \code{comp=NULL}:
     \itemize{
        \item \code{$plot}: Scree plot
        \item \code{$table}: Table of eigenvalues
      }
  \item If \code{comp >= 1}:
     \itemize{
        \item \code{$table}: Table of eigenvalues
        \item \code{$unrotated}: Unrotated factor loading table
        \item \code{$rotated}: Rotated factor loading table
        \item \code{$pcaobj}: PCA object from function \code{prcomp}
     }
}
}
\description{
This function provides results from an standard principal 
    components analysis in preparation for creating PCA perceptual maps.
    While it can be used for traditional, it may not perform correctly.
}
\details{
For best results, the results of the function should be saved to an object. \cr
REQUIRED PACKAGES:
\itemize{
  \item ggplot2
  \item dplyr
}
}
\examples{
#All components with grouping and preference variables
gb.all <- pcaex(pcadata, group="brand", pref="pref")
gb.all$table
gb.all$plot

#Two components with grouping and preference variables
gb.two <- pcaex(pcadata, group="brand", pref="pref", comp=2)
gb.two$table
gb.two$unrotated
gb.two$rotated
}
