% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales01.R
\docType{data}
\name{sales01}
\alias{sales01}
\title{Durable Manufacturer Sales}
\format{
\itemize{
   \item \code{sales01} contains data for 132 months (132 rows and 2 columns)
}
\describe{
  \item{\code{t}}{Time period}
  \item{\code{sales}}{Monthly sales}
}
}
\source{
Unknown
}
\usage{
data(sales01)
}
\description{
Monthly sales data (in $000s) for a durable goods manufacturer from January 2008 through December 2018.
}
\keyword{datasets}
