% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitsample.R
\name{splitsample}
\alias{splitsample}
\title{Split Sample}
\usage{
splitsample(data, outcome, p = 0.75, seed = 4320)
}
\arguments{
\item{data}{The data to be split.}

\item{outcome}{Name of the outcome/dependent variable to be modeled.}

\item{p}{Percent of data to be in the training sample (remainder is in test).
Defaults to 0.75 if excluded.}

\item{seed}{Seed value for random number generator.
Defaults to 4320 if excluded.}
}
\value{
\itemize{ 
  \item \code{train} The training data; will be assigned to global environment.
  \item \code{test} The testing data; will be assigned to the global environment.
}
}
\description{
This function creates a training and test sample based on the
   data and outcome variable provided
}
\details{
REQUIRED PACKAGES:
\itemize{
  \item caret
}
}
\examples{
splitsample(directmktg, buy, 0.6, 123)
}
